requirements.txt
# Core Framework
streamlit==1.28.1
langgraph==0.0.19

# LangChain Ecosystem
langchain==0.1.9
langchain-openai==0.0.8
langchain-community>=0.0.21

# AI/ML
openai==1.3.0
pydantic==2.5.0

# Data & Scientific Computing
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
